requirements.txt
flask
requests
openai
pymongo
python-dotenv